move redirectModal file
since the redirectModal is only used by the 'aboutScreen'
it is moved to the folder containing the components for that screen
