+++ __tests__/SharedComponents.test.jsx
@@ -9,7 +9,7 @@ import { renderWithRedux } from '../__utils__/render';
 
 import Banner from '../src/components/banner/banner';
 import ProgressBar from '../src/components/modal/progressbar';
-import RedirectModal from '../src/components/modal/redirectModal';
+import RedirectModal from '../src/components/about/redirectModal';
 import QuestionnaireModal from '../src/components/modal/questionnaireModal';
 import ScrollIndicatorWrapper from '../src/components/scrollIndicatorWrapper/scrollIndicatorWrapper';
 import Spinner from '../src/components/spinner/spinner';
